iMIRAGE: imputated miRNA Activity from Gene Expression

Function to impute miRNA expression profile from protein coding expression dataset

imirage(train_pcg, train_mir, my_pcg, gene_index, method = "KNN",
  num = 50, target = "none", ...)

Arguments

train_pcg
training protein coding dataset. a numeric matrix with with row names indicating samples, and column names indicating protein coding gene IDs.

train_mir
training miRNA expression dataset. a numeric matrix with row names indicating samples, and column names indicating miRNA IDs.

my_pcg
test protein coding expression dataset. a numeric matrix with row names indicating samples, and column names indicating protein coding gene IDs.

gene_index
either gene name (character) or index (column number) of miRNA to be imputed.

method
method for imputation, either "RF" for random forests, "KNN" for K-nearest neighbor or "SVM" for support vector machines. Uses KNN by default.

num
number of informative protein coding genes to be used in constructing imputation model. Default is 50 genes.

target
"none" (default), "ts.pairs", or dataframe/matrix/list. this argument accepts character strings to indicate the use of all candidate genes as predictors ("none), or use built-in TargetScan miRNA-gene pairs ("ts.pairs"). also accepts a dataframe , matrix or list object containing a column with names of miRNA and a column with the names of target genes.

...
optional parameters that can be passed on to the machine-learning method: RF (randomForest), KNN (knn.reg) or SVM(svm)

Value

a numeric vector with imputed expression levels of the miRNA

Examples

data(iMIRAGE.datasets)
imirage(GA.pcg, GA.mir, HS.pcg, gene_index="hsa-let-7c", method="KNN", num=50)

#>  [1] 2019.3563 1498.5680 1771.0472 2525.3074 3406.9625 1813.6320 3009.6051
#>  [8] 2198.6933 3072.8239 1307.7639 2962.8586 1277.5310 2258.6073 2110.7417
#> [15] 5194.3184 2963.7463 1765.5263 3669.1882 1952.3878 3368.4370 1316.7190
#> [22] 2017.3545 1671.8867 5657.6180 1037.8135 3214.5502 5665.0500 1815.9127
#> [29] 2612.9871 2567.6086 2341.5214 2787.6283 2312.1913 1230.4309 2020.4602
#> [36]  969.7257 1145.1208  958.9258 1043.2435 1181.7525 1092.2694 2585.8381
#> [43] 4762.9124 1914.3931 1659.7709 2286.6204 1590.2248  987.3592 1146.6056
#> [50] 1086.8198

imirage(GA.pcg, GA.mir, HS.pcg, gene_index=25, method="KNN", num=50)

#>  [1]  493.5657  423.0274  425.6345  675.1162  874.6808  425.8818  764.3591
#>  [8]  496.1472  772.2783  349.8015  710.3138  363.3681  570.3462  429.7653
#> [15] 1275.3751  761.7610  429.0182  924.8438  488.8303  889.6338  323.1427
#> [22]  497.5615  341.6435 1649.1141  275.6821  812.9858 1655.2335  459.7502
#> [29]  643.1109  685.3456  550.1771  672.6437  556.0599  339.5751  455.9140
#> [36]  265.0739  304.1595  204.7350  287.0775  300.1527  252.7089  604.5987
#> [43] 1492.3055  508.2830  387.6598  616.8432  425.9734  211.1994  349.9980
#> [50]  280.1517